+++ order-service/src/main/java/com/cg/repository/OrderRepo.java
@@ -1,11 +1,15 @@
 package com.cg.repository;
 
+import java.util.Optional;
+
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 import com.cg.models.OrderDetails;
 @Repository
-public interface OrderRepo extends MongoRepository<OrderDetails, Integer>
+public interface OrderRepo extends MongoRepository<OrderDetails, Long>
 {
-boolean save(int id);
+boolean save(Long id);
+
+//Optional<OrderDetails> findById(Long orderId);
 }
