+++ order-service/src/main/java/com/cg/services/OrderService.java
@@ -9,8 +9,8 @@ public interface OrderService {
 	public OrderDetails addOrder(OrderDetails order);
 	public List<OrderDetails> orderdetails();
 	public OrderDetails updateOrder(OrderDetails update);
-	public void deleteById(int id);
+	public void deleteById(Long id);
 	public Long getSequenceNumber(String sequenceName);
-	public boolean existsById(int id);
+	public boolean existsById(Long id);
 
 }
